% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_counts.R
\name{adjust_counts}
\alias{adjust_counts}
\title{Prepare 1 count series for use by function fit_trend}
\usage{
adjust_counts(
  data,
  category = NULL,
  preferred_field_method = NULL,
  conversion_factor_A2G = NULL,
  ...
)
}
\arguments{
\item{data}{[data frame] data to be analysed (contains at the minimum counts, dates and count precisions)}

\item{category}{[string] species category: MLB, LLB, LD, elephant, or Giraffe}

\item{preferred_field_method}{[character] 'G' for ground or 'A' for aerial.}

\item{conversion_factor_A2G}{[numeric] multiplicative factor to apply to aerial counts to obtain equivalent ground counts.}
}
\value{
a dataframe with the original data and additional fields required by function fit_trend (trend analysis)
}
\description{
To be usable for the estimation of population trend, count data must be 
accompanied by dates collected and information on precision.
Two fields are compulsory:
    - field 'count': contains the original counts
    - field 'year': contains the date each count was taken (given as a year with decimal part e.g. 2019.33).
    
Precision is preferably provided in the form of a 95% confidence interval (CI) by means of 2 fields:
    - field 'lower_bound': the lower boundary of the 95% CI
    - field 'upper_bound': the upper boundary of the 95% CI
It may also be given in the form of a standard deviation, a variance, or a coefficient of variation.  
If the fields 'lower_bound' and 'upper_bound' are both absent, or that at least one contains NA, 
fields 'sd' (standard deviation), 'var' (variance), and 'cv' (coefficient of variation) are examined 
in this order. When one is found valid, a 95% CI is derived assuming a normal distribution. 

If precision is altogether missing, the count must be a total count or a guesstimate and specified as such 
in a field 'stat_method'. Then, a 95% CI interval will be constructed by applying the rules set out 
in .... 
    - field 'stat_method': either T if a total count, S if a sample count, or G if a guesstimate. 
        
By applying the preceding rules, the function fills in the fields 'cinf' and 'csup' of the output dataframe.  
They contain respectively the lower and upper boundaries of the 95% CI. 

If the series mixes aerial and ground counts, a field 'field_method' must be present. 
    - field 'field_method': either A for an aerial count, or G for a ground count
    
As all counts must refer to the same field method, a reference field method must be chosen. It is specified 
via the parameter preferred_field_method. If this parameter is not specified, the reference field method 
depends on the species category. The species category is specified via the parameter 'category'. 
    - 'category' may be MLB for medium-sized light or brown species (20-150kg), 
                        LLB for large light or brown species (>150kg), 
                        LD for large dark species (>150kg), 
                        Giraffe, 
                        Elephant
By default, the preferred method is aerial for LD, Giraffe and Elephant species, ground for MLB and LLB species.

Conversion to the preferred method uses a category-specific coefficient of conversion, which is the 
multiplicative factor to apply to an aerial count to get an equivalent ground count. The conversion 
coefficient may be provided via the parameter 'conversion_factor_A2G'. Alternatively, it is defined internally   
in the package.
}
\examples{

# No example
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}

Roger PRADEL, \email{roger.pradel@cefe.cnrs.fr}
}
