% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{...}
\usage{
check_data(data, conversion_fac = NULL)
}
\arguments{
\item{data}{[data frame] ...}

\item{conversion_fac}{[numeric] ...}
}
\value{
...
}
\description{
...
}
\examples{

# No example
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}

Roger PRADEL, \email{roger.pradel@cefe.cnrs.fr}
}
