% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{conversion_data}
\alias{conversion_data}
\title{Counts conversion data}
\format{
A data frame with 15 rows (African mammals species) and the following
columns:
\describe{
\item{order}{the order of the species}
\item{family}{the family of the species}
\item{species}{the species binomial name}
\item{english}{the species English name}
\item{french}{the species French name}
\item{category}{the detectability category of the species. One of
\code{MLB} for Medium-sized Light and Brown species (20-150kg),
\code{LLB} for Large Light and Brown species (>150kg),
\code{LD} for Large Dark (>150kg), \code{Elephant}, and \code{Giraffe}}
\item{pref_field_method}{the preferred field method of the species. One of
\code{A} for Aerial counts, and \code{G} for Ground counts}
\item{conversion_fact}{the conversion multiplicative factor (corresponding
to the detectability category) used to convert aerial to ground counts}
}
}
\usage{
conversion_data
}
\description{
This dataset contains information to convert individual counts estimated from
a field method to a preferred field method. The field method can be \code{A}
(aerial counts) or \code{G} (ground counts). See \code{\link[=format_data]{format_data()}} for further
information. User can take this dataset as a template to add information for
missing species. Note that only \code{species}, \code{pref_field_method}, and
\code{conversion_fact} are required.
}
\examples{
data("conversion_data")
conversion_data
}
\keyword{datasets}
