% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-funs.R
\name{filter_series}
\alias{filter_series}
\title{Extract the count series corresponding to a location and/or a species in a list}
\usage{
filter_series(data, species = NULL, location = NULL)
}
\arguments{
\item{data}{a named list. The output of function \code{\link[=format_data]{format_data()}}.}

\item{species}{a character string. A species name.}

\item{location}{a character string. A site name.}
}
\value{
A subset of \code{data}, i.e. a named list.
}
\description{
This function identifies the count series relative to a species and/or
a location in a named list like the output of function \code{\link[=format_data]{format_data()}}. If
both species and location are provided, the series of counts of the species
at the specified location is extracted. Otherwise, all series corresponding
to the specified criterion (species or location) are extracted.
}
\examples{
\dontrun{
## Load Garamba dataset ----
file_path <- system.file("extdata", "garamba_survey.csv", 
                         package = "popbayes")
                         
garamba <- read.csv(file = file_path)

## Format dataset ----
garamba_formatted <- popbayes::format_data(garamba)

## Get series names ----
names(garamba_formatted)

## Or...
popbayes::list_series()

## Get data for Alcelaphus buselaphus (at all sites) ----
popbayes::filter_series(garamba_formatted, species = "Alcelaphus buselaphus")

## Get data at Garamba (for all species) ----
popbayes::filter_series(garamba_formatted, location = "Garamba")

## Get data for Alcelaphus buselaphus at Garamba only ----
popbayes::filter_series(garamba_formatted, location = "Garamba",
                        species = "Alcelaphus buselaphus")
}
}
