% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_series.R
\name{find_series}
\alias{find_series}
\title{Subset a series based on a location and/or a species}
\usage{
find_series(data, species = NULL, location = NULL)
}
\arguments{
\item{data}{a named list. The output of the function \code{\link[=format_data]{format_data()}}.}

\item{species}{a character of length 1. A species name.}

\item{location}{a character of length 1. A site name.}
}
\value{
A subset of \code{data}, i.e. a named list.
}
\description{
From the output of the function \code{\link[=format_data]{format_data()}}, this function subsets the
list of series based on a species name and/or a site. It both the species and
the location are provided, the resulting list will have a length of 1 (data
for the species at the location), otherwise the list will have a length >= 1.
}
\examples{
data(garamba)

## Format dataset ----
garamba_formatted <- format_data(data = garamba)

## Get series names ----
names(garamba_formatted)

## Get data for Alcelaphus buselaphus (at all sites) ----
find_series(data = garamba_formatted, species = "Alcelaphus buselaphus")

## Get data for at Garamba (for all species) ----
find_series(data = garamba_formatted, location = "Garamba")

## Get data for Alcelaphus buselaphus at Garamba only ----
find_series(data = garamba_formatted, species = "Alcelaphus buselaphus", 
            location = "Garamba")
}
