% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-funs.R
\name{format_data}
\alias{format_data}
\title{Format individual counts series}
\usage{
format_data(
  data,
  info = NULL,
  year = "year",
  counts = "counts",
  location = "location",
  species = "species",
  stat_method = "stat_method",
  lower_ci = "lower_ci",
  upper_ci = "upper_ci",
  sd = NULL,
  var = NULL,
  cv = NULL,
  field_method = "field_method",
  pref_field_method = NULL,
  conversion_fact = NULL,
  path = ".",
  na_rm = FALSE
)
}
\arguments{
\item{data}{a data frame with at least five columns: \code{location}, \code{species},
\code{year}, \code{counts}, and \code{stat_method}.

The \code{stat_method} field indicates the method used to estimate counts It
can contain: \code{T} (total counts), \code{G} (guesstimate), and/or \code{S} (sampling).

If individuals counts were estimated by \strong{sampling}, additional column(s)
providing a measure of precision is also required (e.g. \code{lower_ci} and
\code{upper_ci}, or \code{sd}, \code{cv}, \code{var}). Precision metrics can be different
between counts. For instance, some sampling counts can have a \code{sd} value
and others \code{lower_ci} and \code{upper_ci}. In that case three columns are
required (\code{lower_ci}, \code{upper_ci}, and \code{sd}). See above section
\strong{Description} for further information on the computation of the 95\%
confident interval of estimates.

If the individuals were counted by different methods, an additional field
\code{field_method} is also required. It can contain: \code{G} (ground counts)
and/or \code{A} (aerial counts). See above section \strong{Description} for further
information on the counts conversion.

Others fields can be present either in \code{data} or \code{info} (see below).}

\item{info}{(optional) a data frame with species in rows and the following
columns: \code{species} (species name), \code{pref_field_method},and
\code{conversion_fact}. See above section \strong{Description} for further
information on these fields.
Default is \code{NULL} (i.e. these informations must be present in \code{data}
if required).}

\item{year}{a character of length 1. The column name in \code{data} of the year.
This column \code{year} must be in the form 1999, 2000, etc. (numeric).
Default is \code{'year'}.}

\item{counts}{a character of length 1. The column name in \code{data} of the
number of individuals. This column must be numerical.
Default is \code{'counts'}.}

\item{location}{a character of length 1. The column name in \code{data} of the
site. This field is used to distinguish counts series from different sites
(if required) and to create an unique series name.
Default is \code{'location'}.}

\item{species}{a character of length 1. The column name in \code{data} (and
in \code{info} if provided) of the species. This field is used to distinguish
counts series for different species (if required) and to create an unique
series name.
Default is \code{'species'}.}

\item{stat_method}{a character of length 1. The column name in \code{data} of
the method used to estimate individuals counts. It can contain \code{'T'}
(total counts), \code{'G'} (guesstimate), and/or \code{'S'} (sampling). If some
counts are coded as \code{'S'}, precision column(s) must also be provided (see
below).
Default is \code{'stat_method'}.}

\item{lower_ci}{(optional) a character of length 1. The column name in \code{data}
of the lower boundary of the 95\% CI of the estimate (i.e. \code{counts}). If
provided the upper boundary of the 95\% CI (argument \code{upper_ci}) must be
also provided. This argument is only required if some counts have been
estimated by a sampling method. But user may prefer use other precision
measures, e.g. standard deviation (argument \code{sd}), variance (\code{var}), or
coefficient of variation (argument \code{cv}).
Default is \code{'lower_ci'}.}

\item{upper_ci}{(optional) a character of length 1. The column name in \code{data}
of the upper boundary of the 95\% CI of the estimate (i.e. \code{counts}). If
provided the lower boundary of the 95\% CI (argument \code{lower_ci}) must be
also provided.
Default is \code{'upper_ci'}.}

\item{sd}{(optional) a character of length 1. The column name in \code{data} of
the standard deviation of the estimate.
Default is \code{NULL}.}

\item{var}{(optional) a character of length 1. The column name in \code{data} of
the variance of the estimate.
Default is \code{NULL}.}

\item{cv}{(optional) a character of length 1. The column name in \code{data} of
the coefficient of variation of the estimate.
Default is \code{NULL}.}

\item{field_method}{(optional) a character of length 1. The column name in
\code{data} of the field method used to count individuals. Counts can be ground
counts (coded as \code{'G'}) or aerial counts (coded as \code{'A'}). This argument
is optional if individuals have been counted by the same method. See above
section \strong{Description} for further information on the counts conversion.
Default is \code{'field_method'}.}

\item{pref_field_method}{(optional) a character of length 1. The column name
in \code{data} of the preferred field method of the species. This argument is
only required is \code{field_method} is not NULL (i.e. individuals have been
counted by different methods). Alternatively this value can be passed in
\code{info}(or internally retrieved if the species is listed in the package).
See above section \strong{Description} for further information on the counts
conversion.
Default is \code{NULL}.}

\item{conversion_fact}{(optional) a character of length 1. The column name
in \code{data} of the counts conversion factor of the species. This argument is
only required is \code{field_method} is not NULL (i.e. individuals have been
counted by different methods). Alternatively this value can be passed in
\code{info} (or internally retrieved if the species is listed in the package).
See above section \strong{Description} for further information on the counts
conversion.
Default is \code{NULL}.}

\item{path}{a character. The directory to save formatted data.
This directory must exist and can be an absolute or a relative path.}

\item{na_rm}{a logical. If \code{TRUE} counts and precision measures (for
sampling counts) with \code{NA} values will be removed.
Default is \code{FALSE} (returns an error to inform user if \code{NA} are detected).}
}
\value{
A n-elements list (where n is the number of counts series). The name
of each element of this list is a combination of location and species.
Each element of the list is a list with the following content:
\itemize{
\item \emph{location} a character. The name of the series site.
\item \emph{species} a character. The name of the series species.
\item \emph{years} a numerical vector. The year sequence of the series.
\item \emph{n_years} an integer. The number of unique years.
\item \emph{stat_methods} a character vector. The different stat methods
of the series.
\item \emph{field_methods} (optional) a character vector. The different
field methods of the series.
\item \emph{pref_field_method} (optional) a character. The preferred
field method of the species (\code{'A'} or \code{'G'}).
\item \emph{conversion_fact} (optional) a numeric. The conversion factor
of the species used to convert counts from a field method to its
preferred field method.
\item \emph{data_original} a data frame. Original data of the series with
renamed columns.
\item \emph{data_converted} a data frame. Data containing computed
boundaries of the 95\% CI (\code{lower_ci_conv} and \code{upper_ci_conv}). If
counts have been obtained by different field methods, contains also
converted counts (\code{counts_conv}) based on the preferred field method and
conversion factor of the species. This data frame will be used by the
function \code{\link[=fit_trend]{fit_trend()}} (not the original dataset).
}

\strong{Note:} Some original series can be discarded if one of these two
conditions is met:
\itemize{
\item the series contains only zero counts;
\item the series contains only a few counts (< 4 years).
}
}
\description{
This function provides an easy way to get individual counts series ready to
be analyzed by the package \code{popbayes}. It must be used prior to all other
functions.

This function formats individual counts series (passed through the argument
\code{data}) by selecting and renaming columns, checking columns format and
content, and removing missing data (if \code{na_rm = TRUE}). It converts the
original data frame into a list of counts series that will be analyzed later
by the function \code{\link[=fit_trend]{fit_trend()}} to estimate population size trend.

To be usable for the estimation of population trend, counts data must be
accompanied by information on precision. The population trend model requires
a 95\% confident interval (CI).

If estimates are total counts or guesstimates, this function will construct
boundaries of the 95\% CI by applying the rules set out in \strong{\emph{???}}.

If counts were estimated by a sampling method user needs to specify a
measure of precision. Precision is preferably provided in the form of a 95\%
CI by means of two fields: \code{lower_ci} and \code{upper_ci}. It may also be given
in the form of a standard deviation (\code{sd}), a variance (\code{var}), or a
coefficient of variation (\code{cv}). If the fields \code{lower_ci} and \code{upper_ci} are
both absent (or NA), fields \code{sd}, \code{var}, and \code{cv} are examined in this order.
When one is found valid (no missing value), a 95\% CI is derived assuming a
normal distribution.
The field \code{stat_method} must be present in the data frame \code{data} to indicate
if counts are \strong{total counts} (\code{'T'}), \strong{sampling} (\code{'S'}), and/or
\strong{guesstimate} (\code{'G'}) and to avoid misinterpretation of counts.

If the series mixes aerial and ground counts, a field \code{field_method} must
also be present and must contain either \code{'A'} (aerial counts), or \code{'G'}
(ground counts). As all counts must refer to the same field method, a
conversion will be performed to homogenize counts. This conversion is based
on a \strong{preferred field method} and a \strong{conversion factor} both specific to
a species/functional group. This conversion factor (a multiplicative factor)
will be apply to an aerial count to get an equivalent ground count (if the
preferred field method is \code{'G'}) or to a ground count to get an equivalent
aerial count (if the preferred field method is \code{'A'}).
These two parameters, named \code{pref_field_method} and \code{conversion_fact}, can
be present in the data frame \code{data} or in the data frame \code{info}.
Alternatively, the package \code{popbayes} provides their values for some
African large mammals.
If the field \code{field_method} is absent in data, counts are assumed to be
obtained with one field method.
}
\examples{
\dontrun{
data("garamba")

## Format dataset ----
garamba_formatted <- popbayes::format_data(garamba)

class(garamba_formatted)
length(garamba_formatted)

## Retrieve counts series names ----
popbayes::list_series()
}
}
