% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series_data.R
\name{get_series_data}
\alias{get_series_data}
\title{Extract original/converted data from a list}
\usage{
get_series_data(data, converted = TRUE)
}
\arguments{
\item{data}{a named list. The output of the function \code{\link[=format_data]{format_data()}}.}

\item{converted}{a logical. If \code{TRUE} (default) extracts converted data,
otherwise returns original data.}
}
\value{
A data frame.
}
\description{
From the output of the function \code{\link[=format_data]{format_data()}}, this function extracts
data frames (for all counts series) containing converted counts
(\code{converted = TRUE}) or original counts (\code{converted = FALSE}). This function
can also be applied on the output of \code{\link[=find_series]{find_series()}}.

The resulting data frame has no particular use in \code{popbayes} but it can be
useful for users.
}
\examples{
data("garamba")

## Format dataset ----
garamba_formatted <- format_data(garamba)

class(garamba_formatted)

## Extract converted counts data ----
converted_data <- get_series_data(garamba_formatted, converted = TRUE)

class(converted_data)

## Extract original counts data ----
original_data <- get_series_data(garamba_formatted, converted = FALSE)

dim(converted_data)
dim(original_data)
dim(garamba)
}
