% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits_plus.R
\name{inits_plus}
\alias{inits_plus}
\title{Generate initial values of relative rate of increase (r)}
\usage{
inits_plus(data, nc)
}
\arguments{
\item{data}{[list] A named list of the data objects passed to the bugs model.}

\item{nc}{[integer] Number of MCMC chains.}
}
\value{
For internal use only.
}
\description{
...
}
\examples{

# No example
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}

Roger PRADEL, \email{roger.pradel@cefe.cnrs.fr}
}
