% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mammals.R
\docType{data}
\name{mammals}
\alias{mammals}
\title{Example dataset for popbayes package}
\format{A data frame with 69 rows and 11 variables:
  - location: protected area name
  - species: species name
  - year: year
  - counts: species counts
  - cinf: CI95% lower bound
  - csup: CI95% upper bound
  - stat_method: S(ample), T(otal) or G(uesstimate)
  - field_method: A(erial) or T(errestrial)
  - cv: coefficient of variation of species count
  - sd: standard deviation of species count
  - var: variance of species count}
\source{
\url{https://www.github.com/frbcesab/popbayes}
}
\usage{
data(mammals)
}
\description{
...
}
\keyword{datasets}
