% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_formula.R
\name{model_formula}
\alias{model_formula}
\title{Write BUGS/JAGS model code}
\usage{
model_formula(jags = TRUE)
}
\arguments{
\item{jags}{[boolean] If TRUE, write model code for JAGS. Of not for OPENBUGS.}
}
\value{
...
}
\description{
...
}
\examples{

model_formula(jags = TRUE)

model_formula(jags = FALSE)
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.com}
}
