% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_series.R
\name{plot_series}
\alias{plot_series}
\title{Plot original vs converted counts (with confident interval)}
\usage{
plot_series(data, series = NULL, title = TRUE, path = ".", save = FALSE)
}
\arguments{
\item{data}{a named list. The output of the function \code{\link[=format_data]{format_data()}}.}

\item{series}{a numeric or a character. Either a series name or a series
index.}

\item{title}{a logical. If \code{TRUE} (default) a title is added at the top-right
of the plot.}

\item{path}{a character. The directory to save the plot (if \code{save = TRUE}).
This directory must exist and can be an absolute or a relative path.}

\item{save}{a logical. If \code{TRUE} (default is \code{FALSE}) the plot is saved in
\code{path}.}
}
\description{
From the output of the function \code{\link[=format_data]{format_data()}}, this function plots a
panel of two figures for one counts series:
\itemize{
\item on the left side, scatter plot overlapping original and converted counts;
\item on the right side, scatter plot of converted counts with boundaries of the
95\% confident interval.
}
}
\examples{
data(garamba)

## Format dataset ----
garamba_formatted <- format_data(data = garamba)

## Get series names ----
names(garamba_formatted)

## Plot for Alcelaphus buselaphus at Garamba ----
plot_series(data = garamba_formatted,
            series = "garamba__alcelaphus_buselaphus", save = FALSE)
            
## alternatively...
plot_series(data = garamba_formatted, series = 1, save = FALSE)
}
