% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot BUGS model preditions}
\usage{
plot_trend(model)
}
\arguments{
\item{model}{[string] a BUGS model obtained after running \code{popbayes::fit_trend()}.}
}
\value{
...
}
\description{
...
}
\examples{

# No example
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}

Roger PRADEL, \email{roger.pradel@cefe.cnrs.fr}
}
