% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_1series.R
\name{prepare_1series}
\alias{prepare_1series}
\title{Prepare 1 count series for use in function fit_trend}
\usage{
prepare_1series(data, category = NULL, preferred_field_method = NULL,
  conversion_fac = NULL)
}
\arguments{
\item{data}{[data frame] data to be analysed (contains at the minimum counts, dates and count precisions)}

\item{category}{[string] species category}

\item{preferred_field_method}{[character] data to be analysed (contains at the minimum counts, dates and count precision)}

\item{conversion_fac}{[numeric] multiplicative factor to apply to ground counts to obtain equivalent aerial counts}
}
\value{
a dataframe with the original data and additional fields required for trend analysis (with function fit_trend)
}
\description{
To be usable for the estimation of population trend, count data must be accompanied by dates collected and 
information on precision. Two fields are compulsory:
    - field 'count': contains the original counts
    - field 'year': contains the date each count was taken (may be given as a fraction of a year).
    
Precision is preferably provided in the form of a 95% confidence interval (CI) by means of 2 fields:
    - field 'lower_bound': the lower boundary of the 95% CI
    - field 'upper_bound': the upper boundary of the 95% CI
It may also be given in the form of a standard deviation, a variance, or a coefficient of variation. In those 
cases, a 95% CI is derived assuming a normal distribution. If the fields 'lower_bound' and 'upper_bound' are 
both absent, or that at least one contains NA, fields 'sd' (standard deviation), 'var' (variance), and 'cv' 
(coefficient of variation) will be sought in this order.

If precision is altogether missing, a 95% CI interval will be constructed by applying the rules set out 
in .... This requires that the statistical method be known. It must be specified in a field 'stat_method' whether 
the count is a total count or a guesstimate.
    - field 'stat_method': either T if a total count, S if a sample count, or G if a guesstimate. 
    If S, a measure of precision is compulsory.
    
By applying the preceding rules, the function creates two fields of the output dataframe, 'cinf' and 'csup', 
for respectively the lower and upper boundaries of the 95%CI. 

If the series mixes aerial and ground counts, a field 'field_method' must be present. 
    - field 'field_method': either A for an aerial count, or G for a ground count
    
As all counts must refer to the same field method, a reference field method must be chosen. It is specified 
via the parameter preferred_field_method. If this parameter is not specified, the reference field method 
depends on the species category. The species category is specified via the parameter 'category'. Alternatively,
a field 'category' may be present: 
    - field 'category': MLB for medium-sized light or brown species (20-150kg), 
                        LLB for large light or brown species (>150kg), 
                        LD for large dark species (>150kg), 
                        Giraffe, 
                        Elephant
The preferred method is aerial for LD, Giraffe and Elephant species, ground for MLB and LLB species.

Conversion to the preferred method uses a category-specific coefficient of conversion, which is the 
multiplicative factor to apply to a ground count to get an equivalent aerial count. The conversion 
coefficient may be provided via the parameter 'conversion_fac'. Alternatively, it is defined internally   
in the package.
}
\examples{

# No example
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}

Roger PRADEL, \email{roger.pradel@cefe.cnrs.fr}
}
