% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugs-funs.R
\name{read_bugs}
\alias{read_bugs}
\title{Import a list of BUGS outputs previously exported}
\usage{
read_bugs(series = NULL, path = ".")
}
\arguments{
\item{series}{a vector of characters. One or several counts series names.
If \code{NULL} (default) BUGS outputs for all counts series will be imported.
Users can run \code{\link[=list_series]{list_series()}} to get the correct spelling of counts series
names.}

\item{path}{a character. The directory in which BUGS outputs have been
saved by the function \code{\link[=fit_trend]{fit_trend()}}.}
}
\value{
A n-elements list (where n is the number of counts series). See
\code{\link[=fit_trend]{fit_trend()}} for further information.
}
\description{
This function imports a list of BUGS outputs previously exported by
\code{\link[=fit_trend]{fit_trend()}}. Users can import one, several, or all models.
}
\examples{
\dontrun{
## Get counts series names ----
popbayes::list_series()

## Import BUGS outputs for one counts series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus")

## Import BUGS outputs for all counts series ----
popbayes::read_bugs()
}
}
