% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rmax_data}
\alias{rmax_data}
\title{Maximum population growth rates}
\format{
A data frame with 15 rows (species) and the following two columns:
\describe{
\item{species}{the species binomial name (same as in \code{conversion_data})}
\item{rmax}{the maximum population growth rate}
}
}
\usage{
rmax_data
}
\description{
This dataset contains for a list of 15 African mammal species the maximum
population growth rate (i.e. the maximum yearly change in log population
size).
}
\examples{
data("rmax_data")
rmax_data
}
\keyword{datasets}
