% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugs-funs.R
\name{bugs_to_df}
\alias{bugs_to_df}
\title{Extract estimated parameters from a list of BUGS outputs}
\usage{
bugs_to_df(data)
}
\arguments{
\item{data}{a named list of BUGS outputs. The output of \code{\link[=fit_trend]{fit_trend()}} or
\code{\link[=read_bugs]{read_bugs()}}.}
}
\value{
A data frame.
}
\description{
From the output of the function \code{\link[=fit_trend]{fit_trend()}} (or \code{\link[=read_bugs]{read_bugs()}}), this
function extracts estimated parameters into a data frame.

The resulting data frame has no particular use in \code{popbayes} but it can be
useful for users.
}
\examples{
\dontrun{
## Get counts series names ----
popbayes::list_series()

## Import BUGS outputs for one counts series ----
bugs <- popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus")

## Extract estimated parameters ----
popbayes::bugs_to_df(bugs)
}
}
