% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugs-funs.R
\name{diagnostic}
\alias{diagnostic}
\title{Check if a BUGS model has converged}
\usage{
diagnostic(data, threshold = 1.1)
}
\arguments{
\item{data}{a named list of BUGS outputs. The output of \code{\link[=fit_trend]{fit_trend()}} or
\code{\link[=read_bugs]{read_bugs()}}.}

\item{threshold}{a numeric.}
}
\description{
From the output of the function \code{\link[=fit_trend]{fit_trend()}} (or \code{\link[=read_bugs]{read_bugs()}}), this
function checks if the estimation of all parameters of one (or several
BUGS models) has converged. This diagnostic is performed by comparing the
\code{Rhat} value of parameters to a \code{threshold} (default is \code{1.1}). If some
\code{Rhat} values are greater than this threshold (no convergence), a message
listing problematic models is displayed.
}
\examples{
\dontrun{
## Get counts series names ----
popbayes::list_series()

## Import BUGS outputs for one counts series ----
bugs <- popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus")

## Check for convergence ----
popbayes::diagnostic(bugs)
}
}
