% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-funs.R
\name{filter_series}
\alias{filter_series}
\title{Subset a counts series list based on location and/or species}
\usage{
filter_series(data, species = NULL, location = NULL)
}
\arguments{
\item{data}{a named list. The output of the function \code{\link[=format_data]{format_data()}}.}

\item{species}{a character of length 1. A species name.}

\item{location}{a character of length 1. A site name.}
}
\value{
A subset of \code{data}, i.e. a named list.
}
\description{
From the output of the function \code{\link[=format_data]{format_data()}}, this function subsets the
list of counts series based on a species name and/or a site. If both the
species and the location are provided, the resulting list will have a length
of 1 (data for the species at the location), otherwise the list will have a
length >= 1.
}
\examples{
\dontrun{
data("garamba")

## Format dataset ----
garamba_formatted <- popbayes::format_data(garamba)

## Get series names ----
names(garamba_formatted)

## Or...
popbayes::list_series()

## Get data for Alcelaphus buselaphus (at all sites) ----
popbayes::filter_series(garamba_formatted, species = "Alcelaphus buselaphus")

## Get data at Garamba (for all species) ----
popbayes::filter_series(garamba_formatted, location = "Garamba")

## Get data for Alcelaphus buselaphus at Garamba only ----
popbayes::filter_series(garamba_formatted, location = "Garamba",
                        species = "Alcelaphus buselaphus")
}
}
