% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend.R
\name{fit_trend}
\alias{fit_trend}
\title{Run BUGS population model}
\usage{
fit_trend(dsname = NULL, MCMCopt = list(ni = 50000, nt = 3, nb = 10000,
  nc = 2), modelopt = list(100, TRUE), jags = TRUE)
}
\arguments{
\item{dsname}{[string] .}

\item{MCMCopt}{[list] .}

\item{modelopt}{[list] .}

\item{jags}{[bolean] If TRUE, write model bugs code for JAGS, otherwise for OpenBUGS.}
}
\value{
For internal use only.
}
\description{
...
}
\examples{

No example
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}

Roger PRADEL, \email{roger.pradel@cefe.cnrs.fr}
}
