% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_trend.R
\name{fit_trend}
\alias{fit_trend}
\title{Run BUGS population model}
\usage{
fit_trend(dsname = NULL, MCMCopt = list(ni = 50000, nt = 3, nb = 10000,
  nc = 2), modelopt = list(100, TRUE), jags = TRUE)
}
\arguments{
\item{dsname}{[string] the R object name containing data.}

\item{MCMCopt}{[list] a list containing the number of iteration (ni), thin factor (nt), length of burn in (nb), i.e. number of iterations to discard at the beginning and the number of chains (nc).}

\item{modelopt}{[list] a list of two vectors: the model smoothing factor and a boolean indicating if parameter r must be limited by rmax.}

\item{jags}{[bolean] If TRUE, write model bugs code for JAGS, otherwise for OpenBUGS.}
}
\value{
...
}
\description{
...
}
\examples{

# No example
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}

Roger PRADEL, \email{roger.pradel@cefe.cnrs.fr}
}
