% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{garamba}
\alias{garamba}
\title{African mammals survey in the Garamba National Park}
\format{
A data frame with 141 rows (counts) and the following 8 columns:
\describe{
\item{location}{the location of the survey (Garamba)}
\item{species}{the binomial name of the species}
\item{year}{the year of the survey}
\item{stat_method}{the method used to estimate individuals counts. One of
\code{T} (total counts), \code{G} (guesstimate), and \code{S}
(sampling counts)}
\item{field_method}{the field method used to collect data. One of
\code{A} (aerial counts), and \code{G} (ground counts)}
\item{counts}{number of individuals}
\item{lower_ci}{lower boundary of the 95\% confident interval of the counts
(only for sampling counts)}
\item{upper_ci}{upper boundary of the 95\% confident interval of the counts
(only for sampling counts)}
}
}
\usage{
garamba
}
\description{
This dataset contains individual counts of 10 African mammals species
in the Garamba National Park (Democratic Republic of the Congo) from 1976 to
2017.
}
\examples{
data("garamba")
head(garamba, 30)
}
\keyword{datasets}
