% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-funs.R
\name{list_series}
\alias{list_series}
\title{Retrieve the count series names}
\usage{
list_series(path)
}
\arguments{
\item{path}{a character string. The directory in which count series have been
saved by the function \code{\link[=format_data]{format_data()}}.}
}
\value{
A vector of count series names (character strings).
}
\description{
This function retrieves the count series names generated by
the function \code{\link[=format_data]{format_data()}}.
}
\examples{
\dontrun{
data("garamba")

## Format dataset ----
garamba_formatted <- format_data(garamba)

## Retrieve counts series names ----
popbayes::list_series()
}
}
