% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_formula.R
\name{model_formula}
\alias{model_formula}
\title{Write bugs model code}
\usage{
model_formula(jags = TRUE)
}
\arguments{
\item{jags}{[boolean] If TRUE, write model code for JAGS. If FALSE, write model code for OPENBUGS.}
}
\value{
For internal use only. No return, write on hard drive a textfile with bugs model code.
}
\description{
...
}
\examples{

model_formula(jags = TRUE)

model_formula(jags = FALSE)
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.com}

Roger PRADEL, \email{roger.pradel@cefe.cnrs.fr}
}
