% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_formula.R
\name{model_formula}
\alias{model_formula}
\title{Write bugs model code}
\usage{
model_formula(jags = TRUE)
}
\arguments{
\item{jags}{[boolean] If TRUE, write model bugs code for JAGS, otherwise for OpenBUGS.}
}
\value{
For internal use only.
No return. Write on hard drive a textfile with bugs model code.
}
\description{
Unlike OpenBugs, JAGS has a 'data' module. Otherwise the code is essentially the same.
The code generated by this function is saved in a text file later read by the bugs program.
 
The code describes a Bayesian model of the counts of one species/site over time. 
The model attempts to reconstitute the trend of the population size while accounting for 
the precision of the individual counts and the demographic potential of the species. 
It also implements a smoothing rule of the relative rates of increase in successive years.
}
\examples{

model_formula(jags = TRUE)   # JAGS formulation
model_formula(jags = FALSE)  # OpenBUGS formulation
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@fondationbiodiversite.fr}

Roger PRADEL, \email{roger.pradel@cefe.cnrs.fr}
}
