% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-funs.R
\name{plot_series}
\alias{plot_series}
\title{Plot original vs converted counts}
\usage{
plot_series(series, title = TRUE, path = ".", path_fig = ".", save = FALSE)
}
\arguments{
\item{series}{a character of length 1. The counts series names (can be
retrieved by running \code{\link[=list_series]{list_series()}}).}

\item{title}{a logical. If \code{TRUE} (default) a title (series name) is added
on the top-right of the plot.}

\item{path}{a character. The directory in which counts series have been
saved by the function \code{\link[=format_data]{format_data()}}.}

\item{path_fig}{a character. The directory to save the plot
(if \code{save = TRUE}). This directory must exist and can be an absolute or a
relative path.}

\item{save}{a logical. If \code{TRUE} (default is \code{FALSE}) the plot is saved in
\code{path_fig}.}
}
\description{
This function plots a panel of two graphics for one counts series
(previously generated by \code{\link[=format_data]{format_data()}}):
\itemize{
\item on the left side, scatter plot overlapping original and converted counts;
\item on the right side, scatter plot of converted counts with boundaries of the
95\% confident interval.
}
}
\examples{
\dontrun{
data("garamba")

## Format dataset ----
garamba_formatted <- popbayes::format_data(data = garamba)

## Get series names ----
popbayes::list_series()

## Plot for Alcelaphus buselaphus at Garamba ----
plot_series("garamba__alcelaphus_buselaphus", save = FALSE)
}
}
