% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popbayes-package.R
\docType{package}
\name{popbayes-package}
\alias{popbayes}
\alias{popbayes-package}
\title{popbayes: Bayesian Model to Estimate Population Trends from Counts Series}
\description{
Infers the trends of one or several populations over time from series of counts. It does so by accounting for count precisions (provided or inferred based on expert knowledge, e.g. for guesstimates), smoothing the population rate of increase over time, and accounting for the maximum demographic potential of species. Inference is carried out in a Bayesian framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/popbayes}
  \item Report bugs at \url{https://github.com/frbcesab/popbayes/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Roger Pradel \email{roger.pradel@cefe.cnrs.fr} (\href{https://orcid.org/0000-0002-2684-9251}{ORCID})
}

}
\keyword{internal}
