% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popbayes-package.R
\docType{package}
\name{popbayes-package}
\alias{popbayes}
\alias{popbayes-package}
\title{popbayes: Title in Title Case}
\description{
Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do 
    eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad 
    minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex 
    ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate 
    velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat 
    cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id 
    est laborum.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/popbayes}
  \item Report bugs at \url{https://github.com/frbcesab/popbayes/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Roger Pradel \email{roger.pradel@cefe.cnrs.fr} (\href{https://orcid.org/0000-0002-2684-9251}{ORCID})
}

}
\keyword{internal}
