% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugs-funs.R
\name{read_bugs}
\alias{read_bugs}
\title{Import a list of BUGS outputs previously exported}
\usage{
read_bugs(series = NULL, path = ".")
}
\arguments{
\item{series}{a vector of character strings. One or several count series
names. If \code{NULL} (default) BUGS outputs for all count series will be
imported. Users can run \code{\link[=list_series]{list_series()}} to get the correct spelling of
count series names.}

\item{path}{a character string. The directory in which BUGS outputs have
been saved by the function \code{\link[=fit_trend]{fit_trend()}}.}
}
\value{
An n-element list (where n is the number of count series). See
\code{\link[=fit_trend]{fit_trend()}} for further information.
}
\description{
This function imports a list of BUGS outputs previously exported by
\code{\link[=fit_trend]{fit_trend()}}. Users can import one, several, or all models.
}
\examples{
\dontrun{
## Get count series names ----
popbayes::list_series()

## Import BUGS outputs for one count series ----
popbayes::read_bugs(series = "garamba__alcelaphus_buselaphus")

## Import BUGS outputs for all count series ----
popbayes::read_bugs()
}
}
