% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-funs.R
\name{series_to_df}
\alias{series_to_df}
\title{Extract original/converted counts series data from a list}
\usage{
series_to_df(data, converted = TRUE)
}
\arguments{
\item{data}{a named list. The output of \code{\link[=format_data]{format_data()}} or \code{\link[=filter_series]{filter_series()}}.}

\item{converted}{a logical. If \code{TRUE} (default) extracts converted counts,
otherwise returns original counts}
}
\value{
A data frame.
}
\description{
From the output of the function \code{\link[=format_data]{format_data()}} (or \code{\link[=filter_series]{filter_series()}}), this
function extracts data frames containing converted counts
(\code{converted = TRUE}) or original counts (\code{converted = FALSE}) for one,
several, or all counts series.

The resulting data frame has no particular use in \code{popbayes} but it can be
useful for users.
}
\examples{
\dontrun{
data("garamba")

## Format dataset ----
garamba_formatted <- popbayes::format_data(garamba)

class(garamba_formatted)

## Extract converted counts data ----
converted_data <- popbayes::series_to_df(garamba_formatted, converted = TRUE)

class(converted_data)

## Extract original counts data ----
original_data <- popbayes::series_to_df(garamba_formatted, converted = FALSE)

dim(converted_data)
dim(original_data)
dim(garamba)
}
}
